Implement JWT and user/pass reusable guards.

+++ backend/src/users/users.controller.ts
@@ -8,11 +8,16 @@ import {
   Param,
   Delete,
   Query,
+  UseGuards,
+  Req,
+  UnauthorizedException,
 } from "@nestjs/common";
 import { UsersService } from "./users.service";
 import { CreateUserDto } from "./dto/create-user.dto";
 import { UpdateUserDto } from "./dto/update-user.dto";
 import { UserFindManyQuery } from "./entities/user.entity";
+import { exclude } from "utils/modelHelpers";
+import { JwtAuthGuard } from "src/auth/guards/jwt-auth.guard";
 
 @Controller("/api/users")
 export class UsersController {
@@ -34,20 +39,36 @@ export class UsersController {
   }
 
   @Get(":id")
-  findOne(@Param("id") userId: string) {
-    return this.usersService.findOne({ userId });
+  async findOne(@Param("id") userId: string) {
+    const user = await this.usersService.findOne({ userId });
+    const userWithoutPassword = exclude(user, "password");
+    return userWithoutPassword;
   }
 
+  @UseGuards(JwtAuthGuard)
   @Patch(":id")
-  update(@Param("id") userId: string, @Body() updateUserDto: UpdateUserDto) {
+  update(
+    @Param("id") userId: string,
+    @Body() updateUserDto: UpdateUserDto,
+    @Req() req
+  ) {
+    if (req.user.userId !== userId) {
+      throw new UnauthorizedException();
+    }
+
     return this.usersService.update({
       where: { userId },
       data: updateUserDto,
     });
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete(":id")
-  remove(@Param("id") userId: string) {
+  remove(@Param("id") userId: string, @Req() req) {
+    if (req.user.userId !== userId) {
+      throw new UnauthorizedException();
+    }
+
     return this.usersService.delete({ userId });
   }
 }
